Fix cart validation and transactions

+++ app/products/models.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Float, ForeignKey, Integer, String, Table
+from sqlalchemy import Column, ForeignKey, Integer, Numeric, String, Table
 from sqlalchemy.orm import relationship
 
 from ..database import db
@@ -27,7 +27,7 @@ class Product(db.Model):
 
     id = Column(Integer, primary_key=True, autoincrement=True)
     title = Column(String(100), nullable=False)
-    price = Column(Integer, nullable=False)
+    price = Column(Numeric(10, 2), nullable=False)
     img_path = Column(String(255), nullable=False)
     description = Column(String(255), nullable=False)
 
@@ -39,7 +39,7 @@ class Option(db.Model):
 
     id = Column(Integer, primary_key=True, autoincrement=True)
     name = Column(String(25), nullable=False)
-    coefficient = Column(Float, nullable=False)
+    coefficient = Column(Numeric(8, 3), nullable=False)
 
     categories = relationship(
         'Category', secondary=category_option_association_table, back_populates='options')
